Improve error handling of Linux commands

+++ WiFiHeatMap/Signal.cs
@@ -5,6 +5,6 @@ namespace WiFiHeatMap
         public string MAC { get; set; }
         public string SSID { get; set; }
         public Frequency Frequency { get; set; }
-        public decimal Strength { get; set; }
+        public short Strength { get; set; }
     }
 }
